Update some whitespace logic in `cd` and `mkdir`.
This was mainly to emulate unix's support for whitespace in filenames,
as long as it's wrapped in quotation marks.

For example:
`mkdir "a b"` creates a directory `a b`.

Additionally, the logic for eliminating trailing slashes was updated.
Previously, a while loop was used and substrings were made every
iteration. This is not efficient for a number of reasons, and so regex
was used in its place.