Modified the filesystem to return FSResult.
FSResult is a new type that is essentially just a wrapper for an
error message. This allows us to return the error message itself,
rather than an integer representing some failed status.

The motivation for using a new type rather than just a string is
for clarity's sake; it may not be immediately apparent that what
was returned is, in fact, a potential error that should be dealt with.

I could see this being changed to a string down the line if necessary,
because adding the additional complexity of a new type has its
natural drawbacks. That said, I really think this is a good idea.