Refactor implementation, add cat command.
Implementation for commands has been abstracted into a separate
file than the terminal component. The motivation for this was to
keep front-end logic separate from command implementation, allowing
for simple addition of new commands.

Additionally, the `cat` command has been added. It will output the
contents of a file. There is no official way to add content to a file
yet.

With the above in mind, there is partial implementation on command
redirection, i.e. transferring command output into a file. This would
likely be the main way to add content to a file, until some major
overhaul to how the terminal is rendered. (I would love to have Vim
support -- I would definitely not implement that in-house, though.)