Overhaul how the terminal is displayed.
An input box used to be used, but this caused issues with wrapping,
and was limited by the fact that it is a box. As a work-around, the
input box is no longer rendered (though still exists) and the text
is itself displayed on the screen using divs and spans.

Some drawbacks to this include typing being *slightly* less responsive.
The difference is neglibile, however, and I personally see no issue with
it.

On the other hand, one core piece of complexity that was added is the
cursor. A new component was created to *simulate* a cursor, and is
rendered at the end of the most recent line. This is a band-aid fix if
anything -- some unexpected behavior is bound to occur from this. (What
happens if you use the arrow keys forward or backward..?)