Fixed word wrap and disabled ligatures.
There was an issue where the terminal text was not wrapping properly.
This was a simple css fix.

There was also another issue where, since the cursor is just a "|"
symbol, if the last character of the line was a "-", it would combine
and form a ligature of "-|" (reverse turnstile iirc?) I just turned off
ligatures for the terminal text. Doubt we would need those anyway.