Refactor line display, add basic color code support
This was a pretty major commit.

**Terminal**
The line system was overhauled. Rather than the terminal history being
represented as a Line[], it now uses two different data
structures; TerminalHistoryEntry[] and a DisplayLine[]. The
TerminalHistoryEntry array keeps track of input/output pairs, their
results, raw in-out, etc, whereas DisplayLine is *entirely* for display. A
DisplayLine contains raw content, and whether it requires a prompt to be
shown.

Splitting up the concept of "terminal history" into two semi-independent
data structures comes with its own complexity, but I believe it was the
right move to make. It allows for better management of previous command
reuslts independent from what needs to be displayed. It allows for
command history that is no longer displayed to be recalled using the
up-arrow. Additionally, and most importantly, it allows for making sure
that commands don't fall off the page if there's too much output.

Down the road, I would like to add scrolling, and I feel like it
wouldn't be terribly difficult given the new refactoring here.

**Color codes**
Basic implementation of color codes. Uses the typical unix \e[<COLOR>m
format. Isn't actually used anywhere yet, but works somewhat. There are
some bugs to work out, but that's tomorrow's problem.