Fix some filesystem bugs, added "A" option to ls
There were some filesystem bugs, mostly related to . and .., which have
since been fixed. Down the line I'd like to figure out how to properly
represent both . and .. as FileSystemNodes, but for now their behavior
is hard-coded to be checked in getChildFile.

The -A option has also been added, which allows for ls to show hidden
directories, but not "." and "..". The -a option has been modified to
now actually, correctly, show "." and "..".