Modify `ls` output to be unformatted on redirect.
`ls` now outputs unformatted, line-separated when redirecting. This aligns
closer with how typical linux systems do it.

Additionally, the `clear` command now sends a special escape sequence
that tells the terminal to clear the history. This is a bit better than
the original solution, which had no implementation for `clear`, but
instead searched for whether the input contained the word "clear".