Modify terminal view, tests for parser-scanner.
The terminal takes up the whole screen when viewed
on a vertical screen. I need to do some more work on
the styling to make sure that it still looks good, but
for now it works.

Also, added unit tests for the scanner. Working on fixing
some stuff in preparation for the full bash parser.